% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdatest-package.R
\docType{package}
\name{fdatest-package}
\alias{fdatest}
\alias{fdatest-package}
\title{fdatest: Interval Wise Testing for Functional Data}
\description{
Implementation of local inferential (domain-selection) procedures for functional data in different frameworks (one- and two-sample tests, ANOVA, linear models). The current version of the package requires functional data evaluated on a common uniform grid or objects of class 'fd' as produced by the 'fda' package and provides the adjusted and unadjusted p-value functions based on nonparametric local tests. The package provides tools for visualization: (i) the p-value heatmap, the plot of the p-value functions and the plot of the functional data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://permaverse.github.io/fdatest/}
  \item \url{https://github.com/permaverse/fdatest/}
  \item Report bugs at \url{https://github.com/permaverse/fdatest/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID}) [contributor]

Authors:
\itemize{
  \item Alessia Pini \email{alessia.pini@unicatt.it}
  \item Simone Vantini \email{simone.vantini@polimi.it}
}

}
\keyword{internal}
