% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdatest2.R
\name{fdatest2}
\alias{fdatest2}
\title{Two population local testing procedures}
\usage{
fdatest2(
  data1,
  data2,
  method,
  mu = 0,
  B = 1000L,
  paired = FALSE,
  dx = NULL,
  alternative = "two.sided",
  recycle = TRUE,
  partition = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data1}{First population's data. Either pointwise evaluations of the
functional data set on a uniform grid, or a \code{fd} object from the
package \code{fda}. If pointwise evaluations are provided, \code{data2} is
a matrix of dimensions \code{c(n1,J)}, with \code{J} evaluations on columns
and \code{n1} units on rows.}

\item{data2}{Second population's data. Either pointwise evaluations of the
functional data set on a uniform grid, or a \code{fd} object from the
package \code{fda}. If pointwise evaluations are provided, \code{data2} is
a matrix of dimensions \code{c(n1,J)}, with \code{J} evaluations on columns
and \code{n2} units on rows.}

\item{method}{A character string specifying the method chosen to adjust the
p-value. Should be one of the following: "\code{Global}", "\code{IWT}",
"\code{TWT}", "\code{PCT}", "\code{FDR}".}

\item{mu}{Either a numeric value or a numeric vector or an object of class
\code{\link[fda:fd]{fda::fd}} specifying the functional mean difference under the null
hypothesis. If \code{mu} is a constant, then a constant function is used. If
\code{mu} is a numeric vector, it must correspond to evaluation of the mean
difference function on the \strong{same} grid that has been used to evaluate the
data samples. Defaults to \code{0}.}

\item{B}{An integer value specifying the number of iterations of the MC
algorithm to evaluate the p-value of the permutation tests. Defaults to
\code{1000L}.}

\item{paired}{A boolean value specifying whether a paired test should be
performed. Defaults to \code{FALSE}.}

\item{dx}{A numeric value specifying the discretization step of the grid used
to evaluate functional data when it is provided as objects of class
\code{\link[fda:fd]{fda::fd}}. Defaults to \code{NULL}, in which case a default value of \code{0.01} is
used which corresponds to a grid of size \code{100L}. Unused if functional data
is provided in the form of matrices.}

\item{alternative}{A string specifying the type of alternative hypothesis.
Choices are \code{"two.sided"}, \code{"less"} or \code{"greater"}. Defaults to
\code{"two.sided"}.}

\item{recycle}{A boolean value specifying whether the recycled version of the
interval-wise testing procedure should be used. Default is \code{TRUE}. See
Pini and Vantini (2017) for details. Defaults to \code{TRUE}.}

\item{partition}{Used only if \code{method}="\code{PCT}". The partition to be
used for PCT procedure. Default is \code{NULL}.}

\item{verbose}{A boolean value specifying whether to print the progress of
the computation. Defaults to \code{FALSE}.}
}
\value{
An object of class \code{fdatest2} containing at least the following
components:
\itemize{
\item \code{test}: a string vector indicating the type of test performed. In this case
equal to \code{"2pop"}.
\item \code{mu}: evaluation on a grid of the functional mean difference under the null
hypothesis (as entered by the user).
\item \code{unadjusted_pval}: evaluation on a grid of the unadjusted p-value function.
\item \code{adjusted_pval}: evaluation on a grid of the adjusted p-value function.
\item \code{data.eval}: evaluation on a grid of the functional data.
\item \code{ord_labels}: vector of labels indicating the group membership of
\code{data.eval}.
}
}
\description{
The function implements local testing procedures for testing mean differences
between two functional populations. Functional data are tested locally and
unadjusted and adjusted p-value functions are provided. The unadjusted
p-value function controls the point-wise error rate. The adjusted p-value
function can be computed according to the following methods:
}
\details{
\itemize{
\item global testing (controlling the FWER weakly)
\item interval-wise testing (controlling the interval-wise error rate)
\item threshold-wise testing (controlling the FWER asymptotically)
\item partition closed testing (controlling the FWER on a partition)
\item functional Benjamini Hochberg (controlling the FDR)
}
}
\examples{
# Performing the TWT for two populations
TWT.result <- fdatest2(
  NASAtemp$paris, NASAtemp$milan, 
  method = "TWT", B = 10L
)

# Plotting the results of the TWT
plot(
  TWT.result, 
  xrange = c(0, 12), 
  main = 'TWT results for testing mean differences'
)

# Selecting the significant components at 5\% level
which(TWT.result$adjusted_pval < 0.05)

# Performing the IWT for two populations
IWT.result <- fdatest2(
  NASAtemp$paris, NASAtemp$milan, 
  method = "IWT", B = 10L
)

# Plotting the results of the IWT
plot(
  IWT.result, 
  xrange = c(0, 12), 
  main = 'IWT results for testing mean differences'
)

# Selecting the significant components at 5\% level
which(IWT.result$adjusted_pval < 0.05)
}
\references{
Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424

A. Pini and S. Vantini (2017). The Interval Testing Procedure: Inference for
Functional Data Controlling the Family Wise Error Rate on Intervals.
Biometrics 73(3): 835–845.

Lundtorp Olsen, N., Pini, A., & Vantini, S. (2021). False discovery rate for
functional data \emph{TEST} 30, 784–809.
}
\seealso{
See also \code{\link[=plot.fdatest2]{plot.fdatest2()}} for plotting the results.
}
