% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globalaov.R
\name{Globalaov}
\alias{Globalaov}
\title{Global testing procedure for testing functional analysis of variance}
\usage{
Globalaov(
  formula,
  dx = NULL,
  B = 1000L,
  method = c("residuals", "responses"),
  stat = c("Integral", "Max")
)
}
\arguments{
\item{formula}{An object of class \code{\link[stats:formula]{stats::formula}} (or one that can be
coerced to that class) specifying the model to be fitted in a symbolic
fashion. The output variable (left-hand side) of the formula can be either
a matrix of dimension \eqn{n \times J} containing the pointwise evaluations
of \eqn{n} functions on the \strong{same} grid of \eqn{J} points, or an object
of class \code{\link[fda:fd]{fda::fd}}.}

\item{dx}{A numeric value specifying the discretization step of the grid used
to evaluate functional data when it is provided as objects of class
\code{\link[fda:fd]{fda::fd}}. Defaults to \code{NULL}, in which case a default value of \code{0.01} is
used which corresponds to a grid of size \code{100L}. Unused if functional data
is provided in the form of matrices.}

\item{B}{An integer value specifying the number of iterations of the MC
algorithm to evaluate the p-value of the permutation tests. Defaults to
\code{1000L}.}

\item{method}{A string specifying the method used to calculate the p-value of
permutation tests. Choices are either \code{"residuals"} which performs
permutation of residuals under the reduced model according to the Freedman
and Lane scheme or \code{"responses"}, which performs permutation of the
responses, according to the Manly scheme. Defaults to \code{"residuals"}.}

\item{stat}{A string specifying the test statistic used for the global test.
Choices are either \code{"Integral"}, in which case the statistic is defined as
the integral of the F-test statistic over the domain, or \code{"Max"}, in which
case the statistic is defined as the maximum of the F-test statistic over
the domain. Defaults to  \code{"Integral"}.}
}
\value{
An object of class \code{IWTaov}. The function \code{summary} is used to
obtain and print a summary of the results. This object is a list containing
the following components:
\itemize{
\item \code{call}: The matched call.
\item \code{design_matrix}: The design matrix of the functional-on-scalar linear
model.
\item \code{unadjusted_pval_F}: Evaluation on a grid of the unadjusted p-value
function of the functional F-test.
\item \code{adjusted_pval_F}: Evaluation on a grid of the adjusted p-value function
of the functional F-test.
\item \code{unadjusted_pval_factors}: Evaluation on a grid of the unadjusted
p-value function of the functional F-tests on each factor of the analysis
of variance (rows).
\item \code{adjusted.pval.factors}: Adjusted p-values of the functional F-tests on
each factor of the analysis of variance (rows) and each basis coefficient
(columns).
\item \code{Global_pval_F}: Global p-value of the overall test F.
\item \code{Global_pval_factors}: Global p-value of test F involving each factor
separately.
\item \code{data.eval}: Evaluation on a fine uniform grid of the functional data
obtained through the basis expansion.
\item \code{coeff.regr.eval}: Evaluation on a fine uniform grid of the functional
regression coefficients.
\item \code{fitted.eval}: Evaluation on a fine uniform grid of the fitted values of
the functional regression.
\item \code{residuals.eval}: Evaluation on a fine uniform grid of the residuals of
the functional regression.
\item \code{R2.eval}: Evaluation on a fine uniform grid of the functional R-squared
of the regression.
}
}
\description{
The function implements the Global Testing procedure for testing mean
differences between several functional populations in a one-way or multi-way
functional analysis of variance framework. Functional data are tested
globally and unadjusted and adjusted p-value functions are provided. The
unadjusted p-value function controls the point-wise error rate. The adjusted
p-value function controls the family-wise error rate weakly. Since this is a
global test, the adjusted p-value function is constant.
}
\examples{
temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the test
Global.result <- Globalaov(temperature ~ groups, B = 1000)

# Summary of the test results
summary(Global.result)

# Plot of the results
layout(1)
plot(Global.result)

# All graphics on the same device
layout(matrix(1:4, nrow = 2, byrow = FALSE))
plot(
  Global.result, 
  main = 'NASA data', 
  plot.adjpval = TRUE, 
  xlab = 'Day', 
  xrange = c(1, 365)
)
}
\references{
\itemize{
\item Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.
\item D. Freedman and D. Lane (1983). A Nonstochastic Interpretation of Reported
Significance Levels. \emph{Journal of Business & Economic Statistics} 1.4,
292-298.
\item B. F. J. Manly (2006). Randomization, \emph{Bootstrap and Monte Carlo
Methods in Biology}. Vol. 70. CRC Press.
}
}
\seealso{
See \code{\link{summary.IWTaov}} for summaries and
\code{\link{plot.IWTaov}} for plotting the results. See
\code{\link{ITPaovbspline}} for a functional analysis of variance test
based on B-spline basis expansion. See also \code{\link{IWTlm}} to fit and
test a functional-on-scalar linear model applying the IWT, and
\code{\link{IWT1}}, \code{\link{IWT2}}  for one-population and
two-population tests.
}
