% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITPimage.R
\name{ITPimage}
\alias{ITPimage}
\title{Heatmap plot of the Interval Testing Procedure results}
\usage{
ITPimage(ITP_result, alpha = 0.05, abscissa_range = c(0, 1), nlevel = 20)
}
\arguments{
\item{ITP_result}{Results of the ITP, as created by
\code{\link{ITP1bspline}}, \code{\link{ITP1fourier}},
\code{\link{ITP2bspline}}, \code{\link{ITP2fourier}}, and
\code{\link{ITP2pafourier}}.}

\item{alpha}{Threshold for the interval-wise error rate used for the
hypothesis test. The default is \code{alpha}=0.05.}

\item{abscissa_range}{Range of the plot abscissa. The default is
\code{c(0,1)}.}

\item{nlevel}{Number of desired color levels for the p-value heatmap. The
default is \code{nlevel=20}.}
}
\description{
Plotting function creating a graphical output of the ITP: the p-value
heat-map, the plot of the corrected p-values, and the plot of the functional
data.
}
\examples{
# Performing the ITP for two populations with the B-spline basis
ITP_result <- ITP2bspline(
  NASAtemp$milan, NASAtemp$paris,
  nknots = 20,
  B = 10L
)

# Plotting the results of the ITP
ITPimage(ITP_result, abscissa_range=c(0,12))

# Selecting the significant components for the radius at 5\% level
which(ITP_result$corrected_pval < 0.05)
}
\references{
A. Pini and S. Vantini (2017). The Interval Testing Procedure: Inference for
Functional Data Controlling the Family Wise Error Rate on Intervals.
Biometrics 73(3): 835–845.
}
