% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ftwosample.R
\name{plot.ftwosample}
\alias{plot.ftwosample}
\alias{autoplot.ftwosample}
\title{Plot for Functional Two-Sample Test Results}
\usage{
\method{autoplot}{ftwosample}(
  object,
  xrange = c(0, 1),
  alpha1 = 0.05,
  alpha2 = 0.01,
  ylabel = "Functional Data",
  title = NULL,
  linewidth = 0.5,
  ...
)

\method{plot}{ftwosample}(
  x,
  xrange = c(0, 1),
  alpha1 = 0.05,
  alpha2 = 0.01,
  ylabel = "Functional Data",
  title = NULL,
  linewidth = 0.5,
  ...
)
}
\arguments{
\item{object, x}{An object of class \code{ftwosample}, usually a result of a call
to \code{\link[=functional_two_sample_test]{functional_two_sample_test()}}, \code{\link[=IWT2]{IWT2()}}, \code{\link[=TWT2]{TWT2()}}, \code{\link[=FDR2]{FDR2()}} or
\code{\link[=PCT2]{PCT2()}}.}

\item{xrange}{A length-2 numeric vector specifying the range of the x-axis
for the plots. Defaults to \code{c(0, 1)}. This should match the domain of the
functional data.}

\item{alpha1}{A numeric value specifying the first level of significance used
to select and display significant effects. Defaults to \code{alpha1 = 0.05}.}

\item{alpha2}{A numeric value specifying the second level of significance
used to select and display significant effects. Defaults to \code{alpha2 = 0.01}.}

\item{ylabel}{A string specifying the label of the y-axis of the functional
data plot. Defaults to \code{"Functional Data"}.}

\item{title}{A string specifying the title of the functional data plot.
Defaults to \code{NULL} in which case no title is displayed.}

\item{linewidth}{A numeric value specifying the width of the line for the
functional data plot. Note that the line width for the adjusted p-value
plot will be twice this value. Defaults to \code{linewidth = 0.5}.}

\item{...}{Other arguments passed to specific methods. Not used in this
function.}
}
\value{
The \code{autoplot.ftwosample()} function creates a ggplot object that
displays the functional data and the adjusted p-values. The significant
intervals at levels \code{alpha1} and \code{alpha2} are highlighted in the plots. The
\code{plot.ftwosample()} function is a wrapper around \code{autoplot.ftwosample()}
that prints the plot directly.
}
\description{
The \code{S3} methods \code{autoplot.ftwosample()} and \code{plot.ftwosample()} are methods
for plotting results of functional two-sample tests. They visualize the
functional data and the adjusted p-values obtained from the testing
procedures for mean comparison of two groups. The plots highlight significant
effects at two levels of significance, \code{alpha1} and \code{alpha2}, using shaded
areas.
}
\examples{
# Performing the TWT for two populations
TWT.result <- functional_two_sample_test(
  NASAtemp$paris, NASAtemp$milan,
  correction = "TWT", B = 10L
)

# Plotting the results of the TWT
plot(
  TWT.result,
  xrange = c(0, 12),
  title = 'TWT results for testing mean differences'
)

# Selecting the significant components at 5\% level
which(TWT.result$adjusted_pval < 0.05)

# Performing the IWT for two populations
IWT.result <- functional_two_sample_test(
  NASAtemp$paris, NASAtemp$milan,
  correction = "IWT", B = 10L
)

# Plotting the results of the IWT
plot(
  IWT.result,
  xrange = c(0, 12),
  title = 'IWT results for testing mean differences'
)

# Selecting the significant components at 5\% level
which(IWT.result$adjusted_pval < 0.05)
}
\references{
Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective
functional analysis of variance for supervised statistical profile monitoring
of signal data. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)} 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., &
Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear
models applied to knee kinematic hop data after injury of the anterior
cruciate ligament. \emph{Scandinavian Journal of Statistics} 45(4),
1036-1061.
}
\seealso{
\code{\link[=IWTimage]{IWTimage()}} for the plot of p-values heatmaps (for IWT).
}
