% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ITP1.R
\name{plot.ITP1}
\alias{plot.ITP1}
\title{Plot the p-values of the ITP1 test}
\usage{
\method{plot}{ITP1}(
  x,
  xrange = c(0, 1),
  alpha1 = 0.05,
  alpha2 = 0.01,
  ylab = "Functional Data",
  main = NULL,
  lwd = 1,
  col = 1,
  pch = 16,
  ylim = range(object$data.eval),
  ...
)
}
\arguments{
\item{x}{An object of class \code{ITP1}.}

\item{xrange}{Range of the x-axis.}

\item{alpha1}{Significance level for the first type of error.}

\item{alpha2}{Significance level for the second type of error.}

\item{ylab}{Label for the y-axis.}

\item{main}{Title of the plot.}

\item{lwd}{Line width.}

\item{col}{Color of the lines.}

\item{pch}{Symbol for the points.}

\item{ylim}{Range of the y-axis.}

\item{...}{Additional arguments to be passed to the \code{matplot} function.}
}
\value{
A plot of the p-values of the ITP1 test.
}
\description{
Plot the p-values of the ITP1 test
}
