% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ITPlm.R
\name{plot.ITPlm}
\alias{plot.ITPlm}
\title{Plot the results of the ITPlm function}
\usage{
\method{plot}{ITPlm}(
  x,
  xrange = c(0, 1),
  alpha1 = 0.05,
  alpha2 = 0.01,
  plot.adjpval = FALSE,
  col = c(1, grDevices::rainbow(dim(x$adjusted.pval.t)[1])),
  ylim = range(x$data.eval),
  ylab = "Functional Data",
  main = NULL,
  lwd = 1,
  pch = 16,
  ...
)
}
\arguments{
\item{x}{Results of the ITPlm function}

\item{xrange}{Range of the plot abscissa. The default is \code{c(0,1)}.}

\item{alpha1}{Threshold for the interval-wise error rate used for the
hypothesis test. The default is \code{alpha1}=0.05.}

\item{alpha2}{Threshold for the interval-wise error rate used for the
hypothesis test. The default is \code{alpha2}=0.01.}

\item{plot.adjpval}{Logical. If \code{TRUE}, the plot of the corrected
p-values is displayed. The default is \code{plot.adjpval=FALSE}.}

\item{col}{Color of the lines in the plot. The default is
\code{col=rainbow(dim(x$adjusted.pval.t)[1])}.}

\item{ylim}{Range of the plot ordinate. The default is \code{range(x$data.eval)}.}

\item{ylab}{Label of the plot ordinate. The default is \code{'Functional Data'}.}

\item{main}{Title of the plot. The default is \code{main=NULL}.}

\item{lwd}{Width of the lines in the plot. The default is \code{lwd=1}.}

\item{pch}{Symbol of the points in the plot. The default is \code{pch=16}.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\description{
Plot the results of the ITPlm function
}
