% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fanova.R
\name{summary.fanova}
\alias{summary.fanova}
\title{Summarizing Functional Analysis of Variance Fits}
\usage{
\method{summary}{fanova}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fanova}, usually, a result of a
call to \code{\link[=functional_anova_test]{functional_anova_test()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No value returned. The function \code{\link[=summary.fanova]{summary.fanova()}} computes and
returns a list of summary statistics of the fitted functional analysis of
variance given in \code{object}, using the component \code{call} from its
arguments, plus:
\itemize{
\item \code{factors}: A \eqn{L \times 1} matrix with columns for the factors of ANOVA,
and corresponding (two-sided) IWT-adjusted minimum p-values of the
corresponding tests of significance (i.e., the minimum p-value over all
\code{p} basis components used to describe functional data).
\item \code{R2}: Range of the functional R-squared.
\item \code{ftest}: IWT-adjusted minimum p-value of functional F-test.
}
}
\description{
\code{summary} method for class \code{fanova}. Function returning a summary
of the results of IWT for the test on a functional analysis of variance:
minimum IWT-adjusted p-values of the F-tests on the whole model and on each
factor are reported.
}
\examples{
temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the IWT
IWT_result <- functional_anova_test(
  temperature ~ groups,
  B = 10L,
  correction = "IWT"
)

# Summary of the IWT results
summary(IWT_result)
}
\references{
Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective
functional analysis of variance for supervised statistical profile monitoring
of signal data. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)} 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., &
Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear
models applied to knee kinematic hop data after injury of the anterior
cruciate ligament. \emph{Scandinavian Journal of Statistics} 45(4),
1036-1061.
}
\seealso{
\code{\link[=IWTimage]{IWTimage()}} for the plot of p-values heatmaps and \code{\link[=plot.fanova]{plot.fanova()}}
for the plot of analysis of variance results.
}
