% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ITPaov.R
\name{summary.ITPaov}
\alias{summary.ITPaov}
\title{S3 specialization of the \code{summary()} function for ITPaov objects}
\usage{
\method{summary}{ITPaov}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ITPaov}.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list with the minimum p-values for the factors, the range of
functional R-squared values, and the minimum p-value for the F-test.
}
\description{
This function is a specialization of the \code{summary()} function for objects of
class \code{ITPaov}. It returns a list with the minimum p-values for the factors,
the range of functional R-squared values, and the minimum p-value for the
F-test.
}
