% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ITPlm.R
\name{summary.ITPlm}
\alias{summary.ITPlm}
\title{S3 specialization of summary function for ITPlm objects}
\usage{
\method{summary}{ITPlm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ITPlm}.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list with the following components:
\itemize{
\item \code{call}: The matched call.
\item \code{ttest}: A data frame with the minimum p-values of the t-tests and the
corresponding significance level.
\item \code{R2}: The R-squared value of the model.
\item \code{ftest}: The p-value of the F-test.
}
}
\description{
S3 specialization of summary function for ITPlm objects
}
