% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-one-sample-test.R
\name{functional_one_sample_test}
\alias{functional_one_sample_test}
\title{Local testing procedures for the functional one-sample test}
\usage{
functional_one_sample_test(
  data,
  correction = c("IWT"),
  mu = 0,
  dx = NULL,
  n_perm = 1000L,
  verbose = FALSE,
  aggregation_strategy = c("integral", "max"),
  recycle = TRUE
)
}
\arguments{
\item{data}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the sample data. If the data is provided within a matrix, it
should be of shape \eqn{n \times J} and it should contain in each row one
of the \eqn{n} functions in the sample and in columns the evaluation of
each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{correction}{A string specifying the correction method to perform the
local functional testing procedure and adjust the p-value function.
Currently only \code{"IWT"} is available.}

\item{mu}{Either a numeric value or a numeric vector or an object of class
\code{\link[fda:fd]{fda::fd}} specifying the functional center of symmetry under the null
hypothesis. If \code{mu} is a constant, then a constant function is used. If
\code{mu} is a numeric vector, it must correspond to evaluations of the center
function on the \strong{same} grid that has been used to evaluate the data
sample. Defaults to \code{0}.}

\item{dx}{A numeric value specifying the step of the uniform grid on which
the data are evaluated. If \code{NULL}, the step is automatically inferred from
the data. Defaults to \code{NULL}.}

\item{n_perm}{An integer value specifying the number of permutations to use
for the local testing procedure. Defaults to \code{1000L}.}

\item{verbose}{A boolean value specifying whether to print the progress of
the computation. Defaults to \code{FALSE}.}

\item{aggregation_strategy}{A string specifying the strategy to aggregate the
point-wise test statistics for the correction procedure. Possible values
are \code{"integral"} and \code{"max"}. Defaults to \code{"integral"}.}

\item{recycle}{A boolean value specifying whether to recycle the test
statistic values across permutations for the IWT procedure. Defaults to
\code{TRUE}.}
}
\value{
An object of class \code{fos} containing the following components:
\itemize{
\item \code{data}: A numeric matrix of shape \eqn{n \times J} containing the
evaluation of the \eqn{n} functions on a uniform grid of size \eqn{J}.
\item \code{mu}: A numeric vector of shape \eqn{J} containing the evaluation of the
functional center of symmetry under the null hypothesis on the same uniform
grid used to evaluate the functional sample.
\item \code{unadjusted_pvalues}: A numeric vector of size \eqn{J} containing the
evaluation of the unadjusted p-value function on the same uniform grid
used to evaluate the functional sample.
\item \code{adjusted_pvalues}: A numeric vector of size \eqn{J} containing the
evaluation of the adjusted p-value function on the same uniform grid used
to evaluate the functional sample.
\item \code{correction_method}: A string containing the correction method used to
compute the adjusted p-value function.
}

Optionally, the list may contain the following component:
\itemize{
\item \code{pvalue_matrix}: A numeric matrix of shape \eqn{p \times p} containing
the p-values of the interval-wise tests. Element \eqn{i, j} contains the
p-value of the test performed on the interval indexed by
\eqn{j, j+1, \dots, j+(p-i)}. Only present if the \code{correction} argument
is set to \code{"IWT"}.
}
}
\description{
The function implements local testing procedures for testing
the center of symmetry of a functional population. Functional data are
tested locally and unadjusted and adjusted p-value functions are provided.
The unadjusted p-value function controls the point-wise error rate. The
adjusted p-value function can be computed according to the following
methods:
\itemize{
\item interval-wise testing (controlling the interval-wise error rate)
}
}
\examples{
# Performing the IWT for one population
IWT_result <- functional_one_sample_test(
  NASAtemp$paris, mu = 4, n_perm = 10L
)

# Plotting the results of the IWT
IWTimage(IWT_result, abscissa_range = c(0, 12))

# Selecting the significant components at 5\% level
which(IWT_result$adjusted_pvalues < 0.05)
}
\references{
For the interval-wise testing procedure:
\itemize{
\item Pini, Alessia, and Simone Vantini. 2016. "The interval testing procedure: a
general framework for inference in functional data analysis." Biometrics 72 (3):
835–845.
\item Pini, Alessia, and Simone Vantini. 2017. "Interval-Wise Testing for Functional
Data." Journal of Nonparametric Statistics 29 (2): 407–24.
}
}
\seealso{
\code{\link[=iwt1]{iwt1()}} for calling directly the IWT test and
\code{\link[=plot.fos]{plot.fos()}} and \code{\link[=autoplot.fos]{autoplot.fos()}} for plotting the results.
}
