% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global2.R
\name{Global2}
\alias{Global2}
\title{Two population Global Testing procedure}
\usage{
Global2(
  data1,
  data2,
  mu = 0,
  dx = NULL,
  B = 1000L,
  paired = FALSE,
  stat = c("Integral", "Max", "Integral_std", "Max_std")
)
}
\arguments{
\item{data1}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the first sample. If the data is provided within a
matrix, it should be of shape \eqn{n_1 \times J} and it should contain in
each row one of the \eqn{n_1} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{data2}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the second sample. If the data is provided within a
matrix, it should be of shape \eqn{n_2 \times J} and it should contain in
each row one of the \eqn{n_2} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{mu}{Either a numeric value or a numeric vector or an object of class
\code{\link[fda:fd]{fda::fd}} specifying the functional mean difference under the null
hypothesis. If \code{mu} is a constant, then a constant function is used. If
\code{mu} is a numeric vector, it must correspond to evaluation of the mean
difference function on the \strong{same} grid that has been used to evaluate the
data samples. Defaults to \code{0}.}

\item{dx}{A numeric value specifying the discretization step of the grid used
to evaluate functional data when it is provided as objects of class
\code{\link[fda:fd]{fda::fd}}. Defaults to \code{NULL}, in which case a default value of \code{0.01} is
used which corresponds to a grid of size \code{100L}. Unused if functional data
is provided in the form of matrices.}

\item{B}{An integer value specifying the number of iterations of the MC
algorithm to evaluate the p-value of the permutation tests. Defaults to
\code{1000L}.}

\item{paired}{A boolean value specifying whether a paired test should be
performed. Defaults to \code{FALSE}.}

\item{stat}{A string speicyfing the test statistic to use. Possible values
are:
\itemize{
\item \code{"Integral"}: Integral of the squared sample mean difference.
\item \code{"Max"}: Maximum of the squared sample mean difference.
\item \code{"Integral_std"}: Integral of the squared t-test statistic.
\item \code{"Max_std"}: Maximum of the squared t-test statistic.
}

Defaults to \code{"Integral"}.}
}
\value{
An object of class \code{fdatest2}, containing the following components:
\itemize{
\item \code{test}: String vector indicating the type of test performed. In this case
equal to \code{"2pop"}.
\item \code{mu}: Evaluation on a grid of the functional mean difference under the
null hypothesis (as entered by the user).
\item \code{unadjusted_pval}: Evaluation on a grid of the unadjusted p-value
function (it is a constant function according to the global testing
procedure).
\item \code{adjusted_pval}: Evaluation on a grid of the adjusted p-value function.
\item \code{data.eval}: Evaluation on a grid of the functional data.
\item \code{ord_labels}: Vector of labels indicating the group membership of
\code{data.eval}.
}
}
\description{
The function implements the Global Testing procedure for testing mean
differences between two functional populations. Functional data are tested
locally and unadjusted and adjusted p-value functions are provided. The
unadjusted p-value function controls the point-wise error rate. The adjusted
p-value function controls the interval-wise error rate.
}
\examples{
# Performing the Global for two populations
Global.result <- Global2(NASAtemp$paris, NASAtemp$milan)

# Plotting the results of the Global
plot(
  Global.result, 
  xrange = c(0, 12), 
  main = 'Global results for testing mean differences'
)

# Selecting the significant components at 5\% level
which(Global.result$adjusted_pval < 0.05)
}
\references{
A. Pini and S. Vantini (2017). The Interval Testing Procedure: Inference for
Functional Data Controlling the Family Wise Error Rate on Intervals.
Biometrics 73(3): 835–845.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424
}
\seealso{
See also \code{\link{IWT2}} for local inference. See
\code{\link{plot.fdatest2}} for plotting the results.
}
