% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-global.R
\name{global2}
\alias{global2}
\alias{Global2}
\title{Two population Global Testing procedure}
\usage{
global2(
  data1,
  data2,
  mu = 0,
  dx = NULL,
  n_perm = 1000L,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  standardize = FALSE,
  verbose = FALSE,
  aggregation_strategy = c("integral", "max")
)

Global2(
  data1,
  data2,
  mu = 0,
  dx = NULL,
  B = 1000L,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  statistic = c("Integral", "Max", "Integral_std", "Max_std"),
  verbose = FALSE
)
}
\arguments{
\item{data1}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the first sample. If the data is provided within a
matrix, it should be of shape \eqn{n_1 \times J} and it should contain in
each row one of the \eqn{n_1} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{data2}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the second sample. If the data is provided within a
matrix, it should be of shape \eqn{n_2 \times J} and it should contain in
each row one of the \eqn{n_2} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{mu}{Either a numeric value or a numeric vector or an object of class
\code{\link[fda:fd]{fda::fd}} specifying the functional mean difference under the null
hypothesis. If \code{mu} is a constant, then a constant function is used. If
\code{mu} is a numeric vector, it must correspond to evaluation of the mean
difference function on the \strong{same} grid that has been used to evaluate the
data samples. Defaults to \code{0}.}

\item{dx}{A numeric value specifying the step of the uniform grid on which
the data are evaluated. If \code{NULL}, the step is automatically inferred from
the data. Defaults to \code{NULL}.}

\item{n_perm}{An integer value specifying the number of permutations to use
for the local testing procedure. Defaults to \code{1000L}.}

\item{paired}{A boolean value specifying whether a paired test should be
performed. Defaults to \code{FALSE}.}

\item{alternative}{A string specifying the type of alternative hypothesis.
Choices are \code{"two.sided"}, \code{"less"} or \code{"greater"}. Defaults to
\code{"two.sided"}.}

\item{standardize}{A boolean value specifying whether to standardize the test
statistic. Defaults to \code{FALSE}.}

\item{verbose}{A boolean value specifying whether to print the progress of
the computation. Defaults to \code{FALSE}.}

\item{aggregation_strategy}{A string specifying the strategy to aggregate the
point-wise test statistics for the correction procedure. Possible values
are \code{"integral"} and \code{"max"}. Defaults to \code{"integral"}.}

\item{B}{An integer value specifying the number of permutations to use
for the local testing procedure. Defaults to \code{1000L}.}

\item{statistic}{A string specifying the test statistic to use. Possible
values are:
\itemize{
\item \code{"Integral"}: Integral of the squared sample mean difference.
\item \code{"Max"}: Maximum of the squared sample mean difference.
\item \code{"Integral_std"}: Integral of the squared t-test statistic.
\item \code{"Max_std"}: Maximum of the squared t-test statistic.
}

Defaults to \code{"Integral"}.}
}
\value{
An object of class \code{fts} containing the following components:
\itemize{
\item \code{data}: A numeric matrix of shape \eqn{n \times J} containing the
evaluation of the \eqn{n = n_1 + n_2} functions on a \strong{common} uniform
grid of size \eqn{p}.
\item \code{group_labels}: An integer vector of size \eqn{n = n_1 + n_2} containing
the group membership of each function.
\item \code{mu}: A numeric vector of shape \eqn{J} containing the evaluation of the
functional mean difference under the null hypothesis on the same uniform
grid used to evaluate the functional samples.
\item \code{unadjusted_pvalues}: A numeric vector of size \eqn{J} containing the
evaluation of the unadjusted p-value function on the \strong{same} uniform grid
used to evaluate the functional samples.
\item \code{adjusted_pvalues}: A numeric vector of size \eqn{J} containing the
evaluation of the adjusted p-value functione on the \strong{same} uniform grid
used to evaluate the functional samples.
\item \code{correction_method}: A string containing the correction method used to
compute the adjusted p-value function.
}

Optionally, the list may contain the following components:
\itemize{
\item \code{global_pvalue}: A numeric value containing the global p-value. Only
present if the \code{correction} argument is set to \code{"Global"}.
\item \code{pvalue_matrix}: A numeric matrix of shape \eqn{p \times p} containing
the p-values of the interval-wise tests. Element \eqn{i, j} contains
the p-value of the test performed on the interval indexed by
\eqn{j, j+1 , \dots, j+(p-i)}. Only present if the \code{correction} argument is
set to \code{"IWT"}.
}
}
\description{
The function implements the Global Testing procedure for testing mean
differences between two functional populations. Functional data are tested
globally and a global p-value and a constant adjusted p-value function are
provided. The adjusted p-value function controls the family-wise error rate
weakly.
}
\examples{
# Performing the Global for two populations
Global_result <- global2(NASAtemp$paris, NASAtemp$milan)

# Plotting the results of the Global
plot(
  Global_result,
  xrange = c(0, 12),
  title = "Global results for testing mean differences"
)

# Selecting the significant components at 5\% level
which(Global_result$adjusted_pvalues < 0.05)
}
\references{
\itemize{
\item Hall, Peter, and Nader Tajvidi. 2002. “Permutation Tests for Equality of
Distributions in High-Dimensional Settings.” Biometrika 89 (2): 359–74.
\item Pini, Alessia, Aymeric Stamm, and Simone Vantini. 2018. “Hotelling’s T2 in
Separable Hilbert Spaces.” Journal of Multivariate Analysis 167: 284–305.
}
}
\seealso{
\code{\link[=iwt2]{iwt2()}}, \code{\link[=twt2]{twt2()}}, \code{\link[=pct2]{pct2()}}, \code{\link[=fdr2]{fdr2()}} for calling directly
one of the other tests, \code{\link[=functional_two_sample_test]{functional_two_sample_test()}} for calling the
interface test and \code{\link[=plot.fts]{plot.fts()}} for plotting the results.
}
