% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITP1bspline.R
\name{ITP1bspline}
\alias{ITP1bspline}
\title{One population Interval Testing Procedure with B-spline basis}
\usage{
ITP1bspline(data, mu = 0, order = 2, nknots = dim(data)[2], B = 1000)
}
\arguments{
\item{data}{Pointwise evaluations of the functional data set on a uniform grid.
It is a matrix of dimensions \code{c(n,J)}, with \code{J} evaluations on columns and \code{n} units on rows.}

\item{mu}{The center of symmetry under the null hypothesis: either a constant
(in this case, a constant function is used) or a \code{J}-dimensional vector containing the evaluations on the
same grid which \code{data} are evaluated. The default is \code{mu=0}.}

\item{order}{Order of the B-spline basis expansion. The default is \code{order=2}.}

\item{nknots}{Number of knots of the B-spline basis expansion. The default is \code{nknots=dim(data)[2]}.}

\item{B}{The number of iterations of the MC algorithm to evaluate the p-values of the permutation tests. The defualt is \code{B=1000}.}
}
\value{
\code{ITP1bspline} returns an object of \code{\link{class}} "\code{ITP1}".
An object of class "\code{ITP1}" is a list containing at least the following components:
\item{basis}{String vector indicating the basis used for the first phase of the algorithm. In this case equal to \code{"B-spline"}.}
\item{test}{String vector indicating the type of test performed. In this case equal to \code{"1pop"}.}
\item{mu}{Center of symmetry under the null hypothesis (as entered by the user).}
\item{coeff}{Matrix of dimensions \code{c(n,p)} of the \code{p} coefficients of the B-spline basis expansion. Rows are associated to units and columns to the basis index.}
\item{pval}{Unadjusted p-values for each basis coefficient.}
\item{pval.matrix}{Matrix of dimensions \code{c(p,p)} of the p-values of the multivariate tests. The element \code{(i,j)} of matrix \code{pval.matrix} contains the p-value of the joint NPC test of the components \code{(j,j+1,...,j+(p-i))}.}
\item{adjusted.pval}{Adjusted p-values for each basis coefficient.}
\item{labels}{Labels indicating the population membership of each data (in this case always equal to \code{1}).}
\item{data.eval}{Evaluation on a fine uniform grid of the functional data obtained through the basis expansion.}
\item{heatmap.matrix}{Heatmap matrix of p-values (used only for plots).}
}
\description{
The function implements the Interval Testing Procedure for testing the center of symmetry of a
functional population evaluated on a uniform grid. Data are represented by means of the B-spline expansion and the
significance of each basis coefficient is tested with an interval-wise control of the Family Wise Error Rate.
The default parameters of the basis expansion lead to the piece-wise interpolating function.
}
\examples{
# Importing the NASA temperatures data set
data(NASAtemp)

# Performing the ITP for one population with the B-spline basis
ITP.result <- ITP1bspline(NASAtemp$paris,mu=4,nknots=50,B=1000)

# Plotting the results of the ITP
plot(ITP.result,xrange=c(0,12),main='Paris temperatures')

# Plotting the p-value heatmap
ITPimage(ITP.result,abscissa.range=c(0,12))

# Selecting the significant components at 5\% level
which(ITP.result$adjusted.pval < 0.05)

}
\references{
A. Pini and S. Vantini (2017).
The Interval Testing Procedure: Inference for Functional Data Controlling the Family Wise Error Rate on Intervals. Biometrics 73(3): 835–845.
}
\seealso{
See also \code{\link{plot.ITP1}} and \code{\link{ITPimage}} for plotting the results,
 \code{\link{ITP1fourier}} for ITP based on Fourier basis, \code{\link{IWT1}} for a one-sample test that is not based on
 an a-priori selected basis expansion.
}
