% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IWT1.R
\name{IWT1}
\alias{IWT1}
\title{One population Interval Wise Testing procedure}
\usage{
IWT1(data, mu = 0, B = 1000, dx = NULL, recycle = TRUE)
}
\arguments{
\item{data}{Either pointwise evaluations of the functional data set on a uniform grid, or a \code{fd} object from the package \code{fda}.
If pointwise evaluations are provided, \code{data} is a matrix of dimensions \code{c(n,J)}, with \code{J} evaluations on columns and \code{n} units on rows.}

\item{mu}{The center of symmetry under the null hypothesis. Three possibilities are available for \code{mu}:
a constant (in this case, a constant function is used);
a \code{J}-dimensional vector containing the evaluations on the same grid which \code{data} are evaluated;
a \code{fd} object from the package \code{fda} containing one function.
The default is \code{mu=0}.}

\item{B}{The number of iterations of the MC algorithm to evaluate the p-values of the permutation tests. The defualt is \code{B=1000}.}

\item{dx}{Used only if a \code{fd} object is provided. In this case, \code{dx} is the size of the discretization step of the grid  used to evaluate functional data.
If set to \code{NULL}, a grid of size 100 is used. Default is \code{NULL}.}

\item{recycle}{Flag used to decide whether the recycled version of the IWT should be used (see Pini and Vantini, 2017 for details). Default is \code{TRUE}.}
}
\value{
\code{IWT1} returns an object of \code{\link{class}} "\code{IWT1}".
An object of class "\code{IWT1}" is a list containing at least the following components:
\item{test}{String vector indicating the type of test performed. In this case equal to \code{"1pop"}.}
\item{mu}{Evaluation on a grid of the center of symmetry under the null hypothesis (as entered by the user).}
\item{unadjusted_pval}{Evaluation on a grid of the unadjusted p-value function.}
\item{pval_matrix}{Matrix of dimensions \code{c(p,p)} of the p-values of the multivariate tests. The element \code{(i,j)} of matrix \code{pval.matrix} contains the p-value of the joint NPC test of the components \code{(j,j+1,...,j+(p-i))}.}
\item{adjusted_pval}{Evaluation on a grid of the adjusted p-value function.}
\item{data.eval}{Evaluation on a grid of the functional data.}
}
\description{
The function implements the Interval Wise Testing procedure for testing the center of symmetry of a
functional population. Functional data are tested locally and  unadjusted  and adjusted p-value
functions are provided. The unadjusted p-value function controls the point-wise error rate. The adjusted p-value function controls the
interval-wise error rate.
}
\examples{
# Importing the NASA temperatures data set
data(NASAtemp)

# Performing the IWT for one population
IWT.result <- IWT1(NASAtemp$paris,mu=4)

# Plotting the results of the IWT
plot(IWT.result,xrange=c(0,12),main='Paris temperatures')

# Plotting the p-value heatmap
IWTimage(IWT.result,abscissa_range=c(0,12))

# Selecting the significant components at 5\% level
which(IWT.result$adjusted_pval < 0.05)

}
\references{
A. Pini and S. Vantini (2017).
The Interval Testing Procedure: Inference for Functional Data Controlling the Family Wise Error Rate on Intervals. Biometrics 73(3): 835–845.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data. \emph{Journal of Nonparametric Statistics}, 29(2), 407-424
}
\seealso{
See also \code{\link{plot.IWT1}} and \code{\link{IWTimage}} for plotting the results.
}
