% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IWT2.R
\name{IWT2}
\alias{IWT2}
\title{Two population Interval Wise Testing procedure}
\usage{
IWT2(data1, data2, mu = 0, B = 1000, paired = FALSE, dx = NULL,
  recycle = TRUE, alternative = "two.sided")
}
\arguments{
\item{data1}{First population's data. Either pointwise evaluations of the functional data set on a uniform grid, or a \code{fd} object from the package \code{fda}.
If pointwise evaluations are provided, \code{data2} is a matrix of dimensions \code{c(n1,J)}, with \code{J} evaluations on columns and \code{n1} units on rows.}

\item{data2}{Second population's data. Either pointwise evaluations of the functional data set on a uniform grid, or a \code{fd} object from the package \code{fda}.
If pointwise evaluations are provided, \code{data2} is a matrix of dimensions \code{c(n1,J)}, with \code{J} evaluations on columns and \code{n2} units on rows.}

\item{mu}{Functional mean difference under the null hypothesis. Three possibilities are available for \code{mu}:
a constant (in this case, a constant function is used);
a \code{J}-dimensional vector containing the evaluations on the same grid which \code{data} are evaluated;
a \code{fd} object from the package \code{fda} containing one function.
The default is \code{mu=0}.}

\item{B}{The number of iterations of the MC algorithm to evaluate the p-values of the permutation tests. The defualt is \code{B=1000}.}

\item{paired}{Flag indicating whether a paired test has to be performed. Default is \code{FALSE}.}

\item{dx}{Used only if a \code{fd} object is provided. In this case, \code{dx} is the size of the discretization step of the grid  used to evaluate functional data.
If set to \code{NULL}, a grid of size 100 is used. Default is \code{NULL}.}

\item{recycle}{Flag used to decide whether the recycled version of the IWT should be used (see Pini and Vantini, 2017 for details). Default is \code{TRUE}.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "\code{two.sided}" (default), "\code{greater}" or "\code{less}".}
}
\value{
\code{IWT2} returns an object of \code{\link{class}} "\code{IWT2}".
An object of class "\code{IWT2}" is a list containing at least the following components:
\item{test}{String vector indicating the type of test performed. In this case equal to \code{"2pop"}.}
\item{mu}{Evaluation on a grid of the functional mean difference under the null hypothesis (as entered by the user).}
\item{unadjusted_pval}{Evaluation on a grid of the unadjusted p-value function.}
\item{pval_matrix}{Matrix of dimensions \code{c(p,p)} of the p-values of the intervalwise tests. The element \code{(i,j)} of matrix \code{pval.matrix} contains the p-value of the test contains the p-value of the test of interval indexed by \code{(j,j+1,...,j+(p-i))}.}
\item{adjusted_pval}{Evaluation on a grid of the adjusted p-value function.}
\item{data.eval}{Evaluation on a grid of the functional data.}
\item{ord_labels}{Vector of labels indicating the group membership of data.eval}
}
\description{
The function implements the Interval Wise Testing procedure for testing mean differences between two
functional populations. Functional data are tested locally and unadjusted and adjusted p-value
functions are provided. The unadjusted p-value function controls the point-wise error rate. The adjusted p-value function controls the
interval-wise error rate.
}
\examples{
# Importing the NASA temperatures data set
data(NASAtemp)

# Performing the IWT for two populations
IWT.result <- IWT2(NASAtemp$paris,NASAtemp$milan)

# Plotting the results of the IWT
plot(IWT.result,xrange=c(0,12),main='IWT results for testing mean differences')

# Plotting the p-value heatmap
IWTimage(IWT.result,abscissa.range=c(0,12))

# Selecting the significant components at 5\% level
which(IWT.result$adjusted_pval < 0.05)

}
\references{
A. Pini and S. Vantini (2017).
The Interval Testing Procedure: Inference for Functional Data Controlling the Family Wise Error Rate on Intervals. Biometrics 73(3): 835–845.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data. \emph{Journal of Nonparametric Statistics}, 29(2), 407-424
}
\seealso{
See also \code{\link{plot.IWT2}} and \code{\link{IWTimage}} for plotting the results.
}
