% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITP2bspline.R, R/ITP2fourier.R,
%   R/ITP2pafourier.R, R/IWT2.R
\name{ITP2bspline}
\alias{ITP2bspline}
\alias{ITP2fourier}
\alias{ITP2pafourier}
\alias{IWT2}
\title{Two population Interval Wise Testing procedure}
\usage{
ITP2bspline(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  order = 2,
  nknots = dim(data1)[2]
)

ITP2fourier(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  maxfrequency = floor(dim(data1)[2]/2)
)

ITP2pafourier(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  maxfrequency = floor(dim(data1)[2]/2)
)

IWT2(
  data1,
  data2,
  mu = 0,
  dx = NULL,
  B = 1000L,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  verbose = FALSE,
  recycle = TRUE
)
}
\arguments{
\item{data1}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the first sample. If the data is provided within a
matrix, it should be of shape \eqn{n_1 \times J} and it should contain in
each row one of the \eqn{n_1} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{data2}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the second sample. If the data is provided within a
matrix, it should be of shape \eqn{n_2 \times J} and it should contain in
each row one of the \eqn{n_2} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{mu}{Either a numeric value or a numeric vector or an object of class
\code{\link[fda:fd]{fda::fd}} specifying the functional mean difference under the null
hypothesis. If \code{mu} is a constant, then a constant function is used. If
\code{mu} is a numeric vector, it must correspond to evaluation of the mean
difference function on the \strong{same} grid that has been used to evaluate the
data samples. Defaults to \code{0}.}

\item{B}{An integer value specifying the number of iterations of the MC
algorithm to evaluate the p-value of the permutation tests. Defaults to
\code{1000L}.}

\item{paired}{A boolean value specifying whether a paired test should be
performed. Defaults to \code{FALSE}.}

\item{order}{Order of the B-spline basis expansion. Defaults to \code{2L}.}

\item{nknots}{Number of knots of the B-spline basis expansion. Defaults to
\code{dim(data1)[2]}.}

\item{maxfrequency}{The maximum frequency to be used in the Fourier basis
expansion of data. Defaults to \code{floor(dim(data1)[2] / 2)}, leading to an
interpolating expansion.}

\item{dx}{A numeric value specifying the discretization step of the grid used
to evaluate functional data when it is provided as objects of class
\code{\link[fda:fd]{fda::fd}}. Defaults to \code{NULL}, in which case a default value of \code{0.01} is
used which corresponds to a grid of size \code{100L}. Unused if functional data
is provided in the form of matrices.}

\item{alternative}{A string specifying the type of alternative hypothesis.
Choices are \code{"two.sided"}, \code{"less"} or \code{"greater"}. Defaults to
\code{"two.sided"}.}

\item{verbose}{A boolean value specifying whether to print the progress of
the computation. Defaults to \code{FALSE}.}

\item{recycle}{A boolean value specifying whether the recycled version of the
interval-wise testing procedure should be used. Default is \code{TRUE}. See
Pini and Vantini (2017) for details. Defaults to \code{TRUE}.}
}
\value{
An object of class \code{\link{IWT2}}, which is a list containing at
least the following components:
\itemize{
\item \code{test}: String vector indicating the type of test performed. In this case
equal to \code{"2pop"}.
\item \code{mu}: Evaluation on a grid of the functional mean difference under the null
hypothesis (as entered by the user).
\item \code{unadjusted_pval}: Evaluation on a grid of the unadjusted p-value function.
\item \code{pval_matrix}: Matrix of dimensions \code{c(p, p)} of the p-values of the
interval-wise tests. The element \verb{(i, j)} of matrix \code{pval.matrix} contains
the p-value of the test contains the p-value of the test of interval indexed
by \verb{(j,j+1,...,j+(p-i))}.
\item \code{adjusted_pval}: Evaluation on a grid of the adjusted p-value function.
\item \code{data.eval}: Evaluation on a grid of the functional data.
\item \code{ord_labels}: Vector of labels indicating the group membership of
\code{data.eval}.
}
}
\description{
The function implements the Interval Wise Testing procedure for testing mean
differences between two functional populations. Functional data are tested
locally and unadjusted and adjusted p-value functions are provided. The
unadjusted p-value function controls the point-wise error rate. The adjusted
p-value function controls the interval-wise error rate.
}
\examples{
# Performing the IWT for two populations
IWT.result <- IWT2(NASAtemp$paris, NASAtemp$milan, B = 10L)

# Plotting the results of the IWT
plot(
  IWT.result, 
  xrange = c(0, 12), 
  main = 'IWT results for testing mean differences'
)

# Plotting the p-value heatmap
IWTimage(IWT.result, abscissa_range = c(0, 12))

# Selecting the significant components at 5\% level
which(IWT.result$adjusted_pval < 0.05)
}
\references{
A. Pini and S. Vantini (2017). The Interval Testing Procedure: Inference
for Functional Data Controlling the Family Wise Error Rate on Intervals.
\emph{Biometrics}, 73(3): 835–845.

A. Pini and S. Vantini (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.
}
\seealso{
See also \code{\link{plot.fdatest2}} and \code{\link{IWTimage}} for
plotting the results.
}
