% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITP2bspline.R, R/ITP2fourier.R,
%   R/ITP2pafourier.R, R/ts-iwt.R
\name{ITP2bspline}
\alias{ITP2bspline}
\alias{ITP2fourier}
\alias{ITP2pafourier}
\alias{iwt2}
\alias{IWT2}
\title{Two population Interval Wise Testing procedure}
\usage{
ITP2bspline(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  order = 2,
  nknots = dim(data1)[2]
)

ITP2fourier(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  maxfrequency = floor(dim(data1)[2]/2)
)

ITP2pafourier(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  maxfrequency = floor(dim(data1)[2]/2)
)

iwt2(
  data1,
  data2,
  mu = 0,
  dx = NULL,
  n_perm = 1000L,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  standardize = FALSE,
  verbose = FALSE,
  aggregation_strategy = c("integral", "max"),
  recycle = TRUE
)

IWT2(
  data1,
  data2,
  mu = 0,
  dx = NULL,
  B = 1000L,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  statistic = c("Integral", "Max", "Integral_std", "Max_std"),
  verbose = FALSE,
  recycle = TRUE
)
}
\arguments{
\item{data1}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the first sample. If the data is provided within a
matrix, it should be of shape \eqn{n_1 \times J} and it should contain in
each row one of the \eqn{n_1} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{data2}{Either a numeric matrix or an object of class \code{\link[fda:fd]{fda::fd}}
specifying the data in the second sample. If the data is provided within a
matrix, it should be of shape \eqn{n_2 \times J} and it should contain in
each row one of the \eqn{n_2} functions in the sample and in columns the
evaluation of each function on a \strong{same} uniform grid of size \eqn{J}.}

\item{mu}{Either a numeric value or a numeric vector or an object of class
\code{\link[fda:fd]{fda::fd}} specifying the functional mean difference under the null
hypothesis. If \code{mu} is a constant, then a constant function is used. If
\code{mu} is a numeric vector, it must correspond to evaluation of the mean
difference function on the \strong{same} grid that has been used to evaluate the
data samples. Defaults to \code{0}.}

\item{B}{An integer value specifying the number of permutations to use
for the local testing procedure. Defaults to \code{1000L}.}

\item{paired}{A boolean value specifying whether a paired test should be
performed. Defaults to \code{FALSE}.}

\item{order}{Order of the B-spline basis expansion. Defaults to \code{2L}.}

\item{nknots}{Number of knots of the B-spline basis expansion. Defaults to
\code{dim(data1)[2]}.}

\item{maxfrequency}{The maximum frequency to be used in the Fourier basis
expansion of data. Defaults to \code{floor(dim(data1)[2] / 2)}, leading to an
interpolating expansion.}

\item{dx}{A numeric value specifying the step of the uniform grid on which
the data are evaluated. If \code{NULL}, the step is automatically inferred from
the data. Defaults to \code{NULL}.}

\item{n_perm}{An integer value specifying the number of permutations to use
for the local testing procedure. Defaults to \code{1000L}.}

\item{alternative}{A string specifying the type of alternative hypothesis.
Choices are \code{"two.sided"}, \code{"less"} or \code{"greater"}. Defaults to
\code{"two.sided"}.}

\item{standardize}{A boolean value specifying whether to standardize the test
statistic. Defaults to \code{FALSE}.}

\item{verbose}{A boolean value specifying whether to print the progress of
the computation. Defaults to \code{FALSE}.}

\item{aggregation_strategy}{A string specifying the strategy to aggregate the
point-wise test statistics for the correction procedure. Possible values
are \code{"integral"} and \code{"max"}. Defaults to \code{"integral"}.}

\item{recycle}{A boolean value specifying whether to recycle the test statistic
values across permutations for the IWT procedure. Defaults to \code{TRUE}.}

\item{statistic}{A string specifying the test statistic to use. Possible
values are:
\itemize{
\item \code{"Integral"}: Integral of the squared sample mean difference.
\item \code{"Max"}: Maximum of the squared sample mean difference.
\item \code{"Integral_std"}: Integral of the squared t-test statistic.
\item \code{"Max_std"}: Maximum of the squared t-test statistic.
}

Defaults to \code{"Integral"}.}
}
\value{
An object of class \code{fts} containing the following components:
\itemize{
\item \code{data}: A numeric matrix of shape \eqn{n \times J} containing the
evaluation of the \eqn{n = n_1 + n_2} functions on a \strong{common} uniform
grid of size \eqn{p}.
\item \code{group_labels}: An integer vector of size \eqn{n = n_1 + n_2} containing
the group membership of each function.
\item \code{mu}: A numeric vector of shape \eqn{J} containing the evaluation of the
functional mean difference under the null hypothesis on the same uniform
grid used to evaluate the functional samples.
\item \code{unadjusted_pvalues}: A numeric vector of size \eqn{J} containing the
evaluation of the unadjusted p-value function on the \strong{same} uniform grid
used to evaluate the functional samples.
\item \code{adjusted_pvalues}: A numeric vector of size \eqn{J} containing the
evaluation of the adjusted p-value functione on the \strong{same} uniform grid
used to evaluate the functional samples.
\item \code{correction_method}: A string containing the correction method used to
compute the adjusted p-value function.
}

Optionally, the list may contain the following components:
\itemize{
\item \code{global_pvalue}: A numeric value containing the global p-value. Only
present if the \code{correction} argument is set to \code{"Global"}.
\item \code{pvalue_matrix}: A numeric matrix of shape \eqn{p \times p} containing
the p-values of the interval-wise tests. Element \eqn{i, j} contains
the p-value of the test performed on the interval indexed by
\eqn{j, j+1 , \dots, j+(p-i)}. Only present if the \code{correction} argument is
set to \code{"IWT"}.
}
}
\description{
The function implements the Interval Wise Testing procedure for testing mean
differences between two functional populations. Functional data are tested
locally and unadjusted and adjusted p-value functions are provided. The
unadjusted p-value function controls the point-wise error rate. The adjusted
p-value function controls the interval-wise error rate.
}
\examples{
# Performing the IWT for two populations
IWT_result <- iwt2(NASAtemp$paris, NASAtemp$milan, n_perm = 10L)

# Plotting the results of the IWT
plot(
  IWT_result,
  xrange = c(0, 12),
  title = "IWT results for testing mean differences"
)

# Plotting the p-value heatmap
IWTimage(IWT_result, abscissa_range = c(0, 12))

# Selecting the significant components at 5\% level
which(IWT_result$adjusted_pvalues < 0.05)
}
\references{
\itemize{
\item Pini, Alessia, and Simone Vantini. 2016. “The interval testing procedure: a
general framework for inference in functional data analysis.” Biometrics 72 (3):
835–845.
\item Pini, Alessia, and Simone Vantini. 2017. “Interval-Wise Testing for Functional
Data.” Journal of Nonparametric Statistics 29 (2): 407–24.
\item Pini, Alessia, Simone Vantini, Bianca Maria Colosimo, and Marco Grasso. 2018.
“Domain-Selective Functional Analysis of Variance for Supervised Statistical
Profile Monitoring of Signal Data.” Journal of the Royal Statistical Society
Series C: Applied Statistics 67 (1): 55–81.
\item Abramowicz, Konrad, Charlotte K Häger, Alessia Pini, Lina Schelin, Sara
Sjöstedt de Luna, and Simone Vantini. 2018. “Nonparametric Inference for
Functional-on-Scalar Linear Models Applied to Knee Kinematic Hop Data After
Injury of the Anterior Cruciate Ligament.” Scandinavian Journal of Statistics 45
(4): 1036–61.
}
}
\seealso{
\code{\link[=global2]{global2()}}, \code{\link[=twt2]{twt2()}}, \code{\link[=pct2]{pct2()}}, \code{\link[=fdr2]{fdr2()}} for calling directly
one of the other tests, \code{\link[=functional_two_sample_test]{functional_two_sample_test()}} for calling the
interface test and \code{\link[=plot.fts]{plot.fts()}} for plotting the results.
}
