% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITP2bspline.R, R/ITP2fourier.R,
%   R/ITP2pafourier.R, R/IWT2.R
\name{ITP2bspline}
\alias{ITP2bspline}
\alias{ITP2fourier}
\alias{ITP2pafourier}
\alias{IWT2}
\title{Two population Interval Wise Testing procedure}
\usage{
ITP2bspline(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  order = 2,
  nknots = dim(data1)[2]
)

ITP2fourier(
  data1,
  data2,
  mu = 0,
  B = 1000,
  paired = FALSE,
  maxfrequency = floor(dim(data1)[2]/2)
)

ITP2pafourier(
  data1,
  data2,
  B = 1000,
  paired = FALSE,
  maxfrequency = floor(dim(data1)[2]/2)
)

IWT2(
  data1,
  data2,
  mu = 0,
  B = 1000,
  dx = NULL,
  recycle = TRUE,
  paired = FALSE,
  alternative = "two.sided"
)
}
\arguments{
\item{data1}{First population's data. Either pointwise evaluations of the
functional data set on a uniform grid, or an \code{\link[fda]{fd}} object.
If pointwise evaluations are provided, it should be a matrix of dimensions
\code{c(n1, J)}, with \code{J} evaluations on columns and \code{n1} units on rows.}

\item{data2}{Second population's data. Either pointwise evaluations of the
functional data set on a uniform grid, or an \code{\link[fda]{fd}} object.
If pointwise evaluations are provided, it should be a matrix of dimensions
\code{c(n2, J)}, with \code{J} evaluations on columns and \code{n2} units on rows.}

\item{mu}{The center of symmetry under the null hypothesis. Three
possibilities are available for \code{mu}:
\itemize{
\item a constant (in this case, a constant function is used);
\item a \code{J}-dimensional vector containing the evaluations on the same grid which
\code{data} are evaluated;
\item a \code{\link[fda]{fd}} object containing one function.
Defaults to \code{0}.
}}

\item{B}{The number of iterations of the MC algorithm to evaluate the p-values of the permutation tests. Defaults to \code{1000L}.}

\item{paired}{Flag indicating whether a paired test has to be performed.
Defaults to \code{FALSE}.}

\item{order}{Order of the B-spline basis expansion. Defaults to \code{2L}.}

\item{nknots}{Number of knots of the B-spline basis expansion. Defaults to
\code{dim(data1)[2]}.}

\item{maxfrequency}{The maximum frequency to be used in the Fourier basis
expansion of data. Defaults to \code{floor(dim(data1)[2] / 2)}, leading to an
interpolating expansion.}

\item{dx}{Used only if an \code{\link[fda]{fd}} object is provided. In this
case, \code{dx} is the size of the discretization step of the gridused to
evaluate functional data. If set to \code{NULL}, a grid of size \code{100L} is used.
Defaults to \code{NULL}.}

\item{recycle}{Flag used to decide whether the recycled version of the IWT
should be used (see Pini and Vantini, 2017 for details). Defaults to
\code{TRUE}.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}
}
\value{
An object of class \code{\link{IWT2}}, which is a list containing at
least the following components:
\itemize{
\item \code{test}: String vector indicating the type of test performed. In this case
equal to \code{"2pop"}.
\item \code{mu}: Evaluation on a grid of the functional mean difference under the null
hypothesis (as entered by the user).
\item \code{unadjusted_pval}: Evaluation on a grid of the unadjusted p-value function.
\item \code{pval_matrix}: Matrix of dimensions \code{c(p, p)} of the p-values of the
interval-wise tests. The element \verb{(i, j)} of matrix \code{pval.matrix} contains
the p-value of the test contains the p-value of the test of interval indexed
by \verb{(j,j+1,...,j+(p-i))}.
\item \code{adjusted_pval}: Evaluation on a grid of the adjusted p-value function.
\item \code{data.eval}: Evaluation on a grid of the functional data.
\item \code{ord_labels}: Vector of labels indicating the group membership of
\code{data.eval}.
}
}
\description{
The function implements the Interval Wise Testing procedure for testing mean
differences between two functional populations. Functional data are tested
locally and unadjusted and adjusted p-value functions are provided. The
unadjusted p-value function controls the point-wise error rate. The adjusted
p-value function controls the interval-wise error rate.
}
\examples{
# Performing the IWT for two populations
IWT.result <- IWT2(NASAtemp$paris, NASAtemp$milan)

# Plotting the results of the IWT
plot(
  IWT.result, 
  xrange = c(0, 12), 
  main = 'IWT results for testing mean differences'
)

# Plotting the p-value heatmap
IWTimage(IWT.result, abscissa_range = c(0, 12))

# Selecting the significant components at 5\% level
which(IWT.result$adjusted_pval < 0.05)

}
\references{
\itemize{
\item A. Pini and S. Vantini (2017). The Interval Testing Procedure: Inference
for Functional Data Controlling the Family Wise Error Rate on Intervals.
\emph{Biometrics}, 73(3): 835–845.
\item A. Pini and S. Vantini (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.
}
}
\seealso{
See also \code{\link{plot.IWT2}} and \code{\link{IWTimage}} for
plotting the results.
}
