% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IWTaov.R
\name{IWTaov}
\alias{IWTaov}
\title{Interval Wise Testing procedure for testing functional analysis of variance}
\usage{
IWTaov(formula, B = 1000L, method = "residuals", dx = NULL, recycle = TRUE)
}
\arguments{
\item{formula}{An object of class "\code{\link{formula}}" (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The output variable of the formula can be either a matrix of dimension
\code{c(n,J)} collecting the pointwise evaluations of \code{n} functional
data on the same grid of \code{J} points, or a \code{fd} object from the
package \code{fda}.}

\item{B}{The number of iterations of the MC algorithm to evaluate the
p-values of the permutation tests. The defualt is \code{B=1000}.}

\item{method}{Permutation method used to calculate the p-value of permutation
tests. Choose "\code{residuals}" for the permutations of residuals under
the reduced model, according to the Freedman and Lane scheme, and
"\code{responses}" for the permutation of the responses, according to the
Manly scheme.}

\item{dx}{Used only if a \code{fd} object is provided. In this case,
\code{dx} is the size of the discretization step of the grid  used to
evaluate functional data. If set to \code{NULL}, a grid of size 100 is
used. Default is \code{NULL}.}

\item{recycle}{Flag used to decide whether the recycled version of the IWT
should be used (see Pini and Vantini, 2017 for details). Default is
\code{TRUE}.}
}
\value{
\code{IWTaov} returns an object of \code{\link{class}}
"\code{IWTaov}". The function \code{summary} is used to obtain and print a
summary of the results. An object of class "\code{IWTaov}" is a list
containing at least the following components:
\itemize{
\item \code{call}: The matched call.
\item \code{design_matrix}: The design matrix of the functional-on-scalar linear
model.
\item \code{unadjusted_pval_F}: Evaluation on a grid of the unadjusted p-value
function of the functional F-test.
\item \code{pval_matrix_F}: Matrix of dimensions \code{c(p,p)} of the p-values of
the intervalwise F-tests. The element \eqn{(i,j)} of matrix \code{pval.matrix}
contains the p-value of the test of interval indexed by
\eqn{(j,j+1,...,j+(p-i))}.
\item \code{adjusted_pval_F}: Evaluation on a grid of the adjusted p-value function
of the functional F-test.
\item \code{unadjusted_pval_factors}: Evaluation on a grid of the unadjusted p-value
function of the functional F-tests on each factor of the analysis of
variance (rows).
\item \code{pval_matrix_factors}: Array of dimensions \code{c(L+1,p,p)} of the p-values
of the multivariate F-tests on factors. The element \eqn{(l,i,j)} of array
\code{pval.matrix} contains the p-value of the joint NPC test on factor \code{l} of
the components \eqn{(j,j+1,...,j+(p-i))}.
\item \code{adjusted_pval_factors}: Adjusted p-values of the functional F-tests on
each factor of the analysis of variance (rows) and each basis coefficient
(columns).
\item \code{data.eval}: Evaluation on a fine uniform grid of the functional data
obtained through the basis expansion.
\item \code{coeff.regr.eval}: Evaluation on a fine uniform grid of the functional
regression coefficients.
\item \code{fitted.eval}: Evaluation on a fine uniform grid of the fitted values of
the functional regression.
\item \code{residuals.eval}: Evaluation on a fine uniform grid of the residuals of
the functional regression.
\item \code{R2.eval}: Evaluation on a fine uniform grid of the functional R-squared
of the regression.
\item \code{heatmap.matrix.F}: Heatmap matrix of p-values of functional F-test (used
only for plots).
\item \code{heatmap.matrix.factors}: Heatmap matrix of p-values of functional
F-tests on each factor of the analysis of variance (used only for plots).
}
}
\description{
The function implements the Interval Wise Testing procedure for testing mean
differences between several functional populations in a one-way or multi-way
functional analysis of variance framework. Functional data are tested locally
and unadjusted and adjusted p-value functions are provided. The unadjusted
p-value function controls the point-wise error rate. The adjusted p-value
function controls the interval-wise error rate.
}
\examples{
temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the IWT
IWT.result <- IWTaov(temperature ~ groups, B = 10L)

# Summary of the ITP results
summary(IWT.result)

# Plot of the IWT results
graphics::layout(1)
plot(IWT.result)

# All graphics on the same device
graphics::layout(matrix(1:4, nrow = 2, byrow = FALSE))
plot(
  IWT.result, 
  main = 'NASA data', 
  plot.adjpval = TRUE, 
  xlab = 'Day', 
  xrange = c(1, 365)
)
}
\references{
Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective
functional analysis of variance for supervised statistical profile monitoring
of signal data. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)} 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., &
Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear
models applied to knee kinematic hop data after injury of the anterior
cruciate ligament. \emph{Scandinavian Journal of Statistics} 45(4),
1036-1061.

D. Freedman and D. Lane (1983). A Nonstochastic Interpretation of Reported
Significance Levels. \emph{Journal of Business & Economic Statistics} 1.4,
292-298.

B. F. J. Manly (2006). Randomization, \emph{Bootstrap and Monte Carlo Methods
in Biology}. Vol. 70. CRC Press.
}
\seealso{
See \code{\link{summary.IWTaov}} for summaries and
\code{\link{plot.IWTaov}} for plotting the results. See
\code{\link{ITPaovbspline}} for a functional analysis of variance test
based on B-spline basis expansion. See also \code{\link{IWTlm}} to fit and
test a functional-on-scalar linear model applying the IWT, and
\code{\link{IWT1}}, \code{\link{IWT2}}  for one-population and
two-population tests.
}
