% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-flm.R
\name{plot.flm}
\alias{plot.flm}
\alias{autoplot.flm}
\title{Plot for Functional Linear Model Test Results}
\usage{
\method{autoplot}{flm}(
  object,
  xrange = c(0, 1),
  alpha1 = 0.05,
  alpha2 = 0.01,
  plot_adjpval = FALSE,
  col = c(1, grDevices::rainbow(dim(object$adjusted_pval_part)[1])),
  ylim = NULL,
  ylabel = "Functional Data",
  title = NULL,
  linewidth = 1,
  type = "l",
  ...
)

\method{plot}{flm}(
  x,
  xrange = c(0, 1),
  alpha1 = 0.05,
  alpha2 = 0.01,
  plot_adjpval = FALSE,
  ylim = NULL,
  col = 1,
  ylab = "Functional Data",
  main = NULL,
  lwd = 0.5,
  type = "l",
  ...
)
}
\arguments{
\item{object, x}{An object of class \code{flm}, usually a result of a call
to \code{\link[=functional_lm_test]{functional_lm_test()}}, \code{\link[=iwt_lm]{iwt_lm()}}, \code{\link[=twt_lm]{twt_lm()}} or \code{\link[=global_lm]{global_lm()}}.}

\item{xrange}{A length-2 numeric vector specifying the range of the x-axis
for the plots. Defaults to \code{c(0, 1)}. This should match the domain of the
functional data.}

\item{alpha1}{A numeric value specifying the first level of significance used
to select and display significant effects. Defaults to \code{alpha1 = 0.05}.}

\item{alpha2}{A numeric value specifying the second level of significance
used to select and display significant effects. Defaults to \code{alpha2 = 0.01}.}

\item{plot_adjpval}{A boolean value specifying whether the plots of adjusted
p-values should be displayed. Defaults to \code{FALSE}.}

\item{col}{An integer specifying the color for the plot of functional data. Defaults
to \code{1L}.}

\item{ylim}{A 2-length numeric vector specifying the range of the y-axis.
Defaults to \code{NULL}, which determines automatically the range from functional data.}

\item{ylabel, ylab}{A string specifying the label of the y-axis of the functional
data plot. Defaults to \code{"Functional Data"}.}

\item{title, main}{A string specifying the title of the functional data plot.
Defaults to \code{NULL} in which case no title is displayed.}

\item{linewidth, lwd}{A numeric value specifying the width of the line for the
functional data plot. Note that the line width for the adjusted p-value
plot will be twice this value. Defaults to \code{0.5}.}

\item{type}{A string specifying the type of plot for the functional data. Defaults
to \code{"l"} for lines.}

\item{...}{Other arguments passed to specific methods. Not used in this
function.}
}
\value{
The \code{autoplot.flm()} function creates a ggplot object that
displays the functional data and the adjusted p-values. The significant
intervals at levels \code{alpha1} and \code{alpha2} are highlighted in the plots.
The \code{plot.flm()} function is a wrapper around \code{autoplot.flm()}
that prints the plot directly.
}
\description{
The \code{S3} methods \code{autoplot.flm()} and \code{plot.flm()} are methods
for plotting results of functional-on-scalar linear model tests. They
visualize the functional regression coefficients and the adjusted p-values
obtained from the testing procedures. The plots highlight significant
effects at two levels of significance, \code{alpha1} and \code{alpha2}, using shaded
areas.
}
\examples{
temperature <- rbind(NASAtemp$milan[, 1:100], NASAtemp$paris[, 1:100])
groups <- c(rep(0, 22), rep(1, 22))

# Performing the IWT
IWT_result <- IWTlm(temperature ~ groups, B = 2L)

# Summary of the IWT results
summary(IWT_result)

# Plot of the IWT results
plot(
  IWT_result,
  main = "NASA data",
  plot_adjpval = TRUE,
  xlab = "Day",
  xrange = c(1, 365)
)
}
\references{
Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective
functional analysis of variance for supervised statistical profile monitoring
of signal data. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)} 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., &
Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear
models applied to knee kinematic hop data after injury of the anterior
cruciate ligament. \emph{Scandinavian Journal of Statistics} 45(4),
1036-1061.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.
}
\seealso{
\code{\link[=IWTimage]{IWTimage()}} for the plot of p-values heatmaps (for IWT).
}
