% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.faov.R
\name{summary.faov}
\alias{summary.faov}
\title{Summarizing Functional Analysis of Variance Fits}
\usage{
\method{summary}{faov}(object, ...)
}
\arguments{
\item{object}{An object of class \code{faov}, usually, a result of a
call to \code{\link[=functional_anova_test]{functional_anova_test()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of summary statistics of the fitted functional analysis of
variance given in \code{object}, using the component \code{call} from its arguments,
plus:
\itemize{
\item \code{factors}: A \eqn{L \times 2} data frame with one row per factor
reporting the minimum adjusted p-value of the corresponding F-test and a
significance code.
\item \code{R2}: A \eqn{2 \times 1} matrix giving the range of the functional
R-squared.
\item \code{ftest}: A \eqn{1 \times 2} data frame reporting the minimum adjusted
p-value of the global F-test and a significance code.
}
}
\description{
\code{summary} method for class \code{faov}. Returns a summary of the results of the
local testing procedure for functional analysis of variance: the minimum
adjusted p-values of the F-test on the whole model and on each factor are
reported.
}
\examples{
temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the IWT
IWT_result <- functional_anova_test(
  temperature ~ groups,
  B = 10L,
  correction = "IWT"
)

# Summary of the IWT results
summary(IWT_result)
}
\references{
Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective
functional analysis of variance for supervised statistical profile monitoring
of signal data. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)} 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., &
Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear
models applied to knee kinematic hop data after injury of the anterior
cruciate ligament. \emph{Scandinavian Journal of Statistics} 45(4),
1036-1061.
}
\seealso{
\code{\link[=IWTimage]{IWTimage()}} for the plot of p-value heatmaps and \code{\link[=plot.faov]{plot.faov()}}
for the plot of analysis of variance results.
}
