% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flm.R
\name{summary.flm}
\alias{summary.flm}
\title{Summarizing Functional Linear Model Fits}
\usage{
\method{summary}{flm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{flm}, usually, a result of a
call to \code{\link[=functional_lm_test]{functional_lm_test()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of summary statistics of the fitted functional linear model
given in \code{object}, using the component \code{call} from its arguments, plus:
\itemize{
\item \code{ttest}: A \eqn{(L+1) \times 2} data frame with one row per model term
(intercept plus each predictor) reporting the minimum adjusted p-value of
the corresponding t-test and a significance code.
\item \code{R2}: A \eqn{2 \times 1} matrix giving the range of the functional
R-squared.
\item \code{ftest}: A \eqn{1 \times 2} data frame reporting the minimum adjusted
p-value of the functional F-test and a significance code.
}
}
\description{
\code{summary} method for class \code{flm}. Returns a summary of the results of the
local testing procedure for a functional-on-scalar linear model: the minimum
adjusted p-values of the F-test on the whole model and the t-tests on each
covariate are reported.
}
\examples{
temperature <- rbind(NASAtemp$milan[, 1:100], NASAtemp$paris[, 1:100])
groups <- c(rep(0, 22), rep(1, 22))

# Performing the IWT
IWT_result <- functional_lm_test(
  temperature ~ groups,
  B = 2L,
  correction = "IWT"
)

# Summary of the IWT results
summary(IWT_result)
}
\references{
Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.

Pini, A., & Vantini, S. (2017). Interval-wise testing for functional data.
\emph{Journal of Nonparametric Statistics}, 29(2), 407-424.

Pini, A., Vantini, S., Colosimo, B. M., & Grasso, M. (2018). Domain‐selective
functional analysis of variance for supervised statistical profile monitoring
of signal data. \emph{Journal of the Royal Statistical Society: Series C
(Applied Statistics)} 67(1), 55-81.

Abramowicz, K., Hager, C. K., Pini, A., Schelin, L., Sjostedt de Luna, S., &
Vantini, S. (2018). Nonparametric inference for functional‐on‐scalar linear
models applied to knee kinematic hop data after injury of the anterior
cruciate ligament. \emph{Scandinavian Journal of Statistics} 45(4),
1036-1061.
}
\seealso{
\code{\link[=IWTimage]{IWTimage()}} for the plot of p-value heatmaps and \code{\link[=plot.flm]{plot.flm()}}
for the plot of functional linear model results.
}
