% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-twt.R
\name{TWTlm}
\alias{TWTlm}
\alias{twt_lm}
\title{Threshold-wise testing procedure for testing functional-on-scalar linear
models}
\usage{
TWTlm(formula, dx = NULL, B = 1000L, method = c("residuals", "responses"))

twt_lm(
  formula,
  dx = NULL,
  n_perm = 1000L,
  method = c("residuals", "responses")
)
}
\arguments{
\item{formula}{An object of class \code{\link[stats:formula]{stats::formula}} (or one that can be
coerced to that class) specifying the model to be fitted in a symbolic
fashion. The output variable (left-hand side) of the formula can be either
a matrix of dimension \eqn{n \times J} containing the pointwise evaluations
of \eqn{n} functions on the \strong{same} grid of \eqn{J} points, or an object
of class \code{\link[fda:fd]{fda::fd}}.}

\item{dx}{A numeric value specifying the discretization step of the grid used
to evaluate functional data when it is provided as objects of class
\code{\link[fda:fd]{fda::fd}}. Defaults to \code{NULL}, in which case a default value of \code{0.01} is
used which corresponds to a grid of size \code{100L}. Unused if functional data
is provided in the form of matrices.}

\item{B}{An integer value specifying the number of iterations of the MC
algorithm to evaluate the p-value of the permutation tests. Defaults to
\code{1000L}.}

\item{method}{A string specifying the method used to calculate the p-value of
permutation tests. Choices are either \code{"residuals"} which performs
permutation of residuals under the reduced model according to the Freedman
and Lane scheme or \code{"responses"}, which performs permutation of the
responses, according to the Manly scheme. Defaults to \code{"residuals"}.}

\item{n_perm}{An integer value specifying the number of permutations for the
permutation tests. Defaults to \code{1000L}.}
}
\value{
An object of class \code{flm} containing the following components:
\itemize{
\item \code{call}: The matched call.
\item \code{design_matrix}: The design matrix of the functional-on-scalar linear
model.
\item \code{unadjusted_pval_F}: Evaluation on a grid of the unadjusted p-value
function of the functional F-test.
\item \code{adjusted_pval_F}: Evaluation on a grid of the adjusted p-value function
of the functional F-test.
\item \code{unadjusted_pval_part}: Evaluation on a grid of the unadjusted p-value
function of the functional F-tests on each factor of the analysis of
variance (rows).
\item \code{adjusted_pval_part}: Adjusted p-values of the functional F-tests on
each factor of the analysis of variance (rows) and each basis coefficient
(columns).
\item \code{data_eval}: Evaluation on a fine uniform grid of the functional data
obtained through the basis expansion.
\item \code{coeff_regr_eval}: Evaluation on a fine uniform grid of the functional
regression coefficients.
\item \code{fitted_eval}: Evaluation on a fine uniform grid of the fitted values of
the functional regression.
\item \code{residuals_eval}: Evaluation on a fine uniform grid of the residuals of
the functional regression.
\item \code{R2_eval}: Evaluation on a fine uniform grid of the functional R-squared
of the regression.
}

Optionally, the list may contain the following components:
\itemize{
\item \code{pval_matrix_F}: Matrix of dimensions \code{c(p,p)} of the p-values of
the intervalwise F-tests. The element \eqn{(i,j)} of matrix \code{pval_matrix}
contains the p-value of the test of interval indexed by
\eqn{(j,j+1,...,j+(p-i))}; this component is present only if \code{correction}
is set to \code{"IWT"}.
\item \code{pval_matrix_part}: Array of dimensions \code{c(L+1,p,p)} of the p-values
of the multivariate F-tests on factors. The element \eqn{(l,i,j)} of array
\code{pval_matrix_part} contains the p-value of the joint NPC test on factor \code{l} of
the components \eqn{(j,j+1,...,j+(p-i))}; this component is present only if
\code{correction} is set to \code{"IWT"}.
\item \code{Global_pval_F}: Global p-value of the overall test F; this component is
present only if \code{correction} is set to \code{"Global"}.
\item \code{Global_pval_part}: Global p-value of test F involving each factor
separately; this component is present only if \code{correction} is set to \code{"Global"}.
}
}
\description{
The function is used to fit and test functional linear models. It can be used
to carry out regression, and analysis of variance. It implements the
Threshold-wise testing procedure (TWT) for testing the significance of the
effects of scalar covariates on a functional population.
}
\examples{
# Defining the covariates
temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the TWT
TWT_result <- TWTlm(temperature ~ groups, B = 100L)
# Summary of the TWT results
summary(TWT_result)

# Plot of the TWT results
plot(
  TWT_result,
  main = 'NASA data',
  plot_adjpval = TRUE,
  xlab = 'Day',
  xrange = c(1, 365)
)

plot(
  TWT_result,
  main = 'NASA data',
  plot_adjpval = TRUE,
  xlab = 'Day',
  xrange = c(1, 365)
)
}
\references{
Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.

D. Freedman and D. Lane (1983). A Nonstochastic Interpretation of Reported
Significance Levels. \emph{Journal of Business & Economic Statistics} 1(4),
292-298.

B. F. J. Manly (2006). Randomization, \emph{Bootstrap and Monte Carlo Methods
in Biology}. Vol. 70. CRC Press.
}
\seealso{
See also \code{\link[=plot.flm]{plot.flm()}} for plotting the results and
\code{\link[=summary.flm]{summary.flm()}} for summarizing the results of the functional analysis
of variance.
}
