% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-twt.R
\name{TWTlm}
\alias{TWTlm}
\title{Threshold-wise testing procedure for testing functional-on-scalar linear
models}
\usage{
TWTlm(formula, dx = NULL, B = 1000L, method = c("residuals", "responses"))
}
\arguments{
\item{formula}{An object of class \code{\link[stats:formula]{stats::formula}} (or one that can be
coerced to that class) specifying the model to be fitted in a symbolic
fashion. The output variable (left-hand side) of the formula can be either
a matrix of dimension \eqn{n \times J} containing the pointwise evaluations
of \eqn{n} functions on the \strong{same} grid of \eqn{J} points, or an object
of class \code{\link[fda:fd]{fda::fd}}.}

\item{dx}{A numeric value specifying the discretization step of the grid used
to evaluate functional data when it is provided as objects of class
\code{\link[fda:fd]{fda::fd}}. Defaults to \code{NULL}, in which case a default value of \code{0.01} is
used which corresponds to a grid of size \code{100L}. Unused if functional data
is provided in the form of matrices.}

\item{B}{An integer value specifying the number of iterations of the MC
algorithm to evaluate the p-value of the permutation tests. Defaults to
\code{1000L}.}

\item{method}{A string specifying the method used to calculate the p-value of
permutation tests. Choices are either \code{"residuals"} which performs
permutation of residuals under the reduced model according to the Freedman
and Lane scheme or \code{"responses"}, which performs permutation of the
responses, according to the Manly scheme. Defaults to \code{"residuals"}.}
}
\value{
An object of class \code{TWTlm}. The function \code{summary} is used to
obtain and print a summary of the results. An object of class
"\code{TWTlm}" is a list containing at least the following components:
\itemize{
\item \code{call}: Call of the function.
\item \code{design_matrix}: Design matrix of the linear model.
\item \code{unadjusted_pval_F}: Unadjusted p-value function of the F test.
\item \code{adjusted_pval_F}: Adjusted p-value function of the F test.
\item \code{unadjusted_pval_part}: Unadjusted p-value functions of the functional
t-tests on each covariate, separately (rows) on each domain point
(columns).
\item \code{adjusted_pval_part}: Adjusted p-values of the functional t-tests on each
covariate (rows) on each domain point (columns).
\item \code{data.eval}: Evaluation of functional data.
\item \code{coeff.regr.eval}: Evaluation of the regression coefficients.
\item \code{fitted.eval}: Evaluation of the fitted values.
\item \code{residuals.eval}: Evaluation of the residuals.
\item \code{R2.eval}: Evaluation of the functional R-squared.
}
}
\description{
The function is used to fit and test functional linear models. It can be used
to carry out regression, and analysis of variance. It implements the
Threshold-wise testing procedure (TWT) for testing the significance of the
effects of scalar covariates on a functional population.
}
\examples{
# Defining the covariates
temperature <- rbind(NASAtemp$milan, NASAtemp$paris)
groups <- c(rep(0, 22), rep(1, 22))

# Performing the TWT
TWT.result <- TWTlm(temperature ~ groups, B = 100L)
# Summary of the TWT results
summary(TWT.result)

# Plot of the TWT results
layout(1)
plot(
  TWT.result,
  main = 'NASA data',
  plot_adjpval = TRUE,
  xlab = 'Day',
  xrange = c(1, 365)
)

# All graphics on the same device
layout(matrix(1:6, nrow = 3, byrow = FALSE))
plot(
  TWT.result,
  main = 'NASA data',
  plot_adjpval = TRUE,
  xlab = 'Day',
  xrange = c(1, 365)
)
}
\references{
Abramowicz, K., Pini, A., Schelin, L., Stamm, A., & Vantini, S. (2022).
“Domain selection and familywise error rate for functional data: A unified
framework. \emph{Biometrics} 79(2), 1119-1132.

D. Freedman and D. Lane (1983). A Nonstochastic Interpretation of Reported
Significance Levels. \emph{Journal of Business & Economic Statistics} 1(4),
292-298.

B. F. J. Manly (2006). Randomization, \emph{Bootstrap and Monte Carlo Methods
in Biology}. Vol. 70. CRC Press.
}
\seealso{
See \code{\link{summary.TWTlm}} for summaries and
\code{\link{plot.TWTlm}} for plotting the results. See also
\code{\link{TWTaov}} to fit and test a functional analysis of variance
applying the TWT, and  \code{\link{TWT2}} for two-population test.
}
